inlined code instrumentation and addressed PR comments
